+++ tests/app/s3_client/test_s3_logo_client.py
@@ -1,16 +1,15 @@
 from collections import namedtuple
-from unittest.mock import Mock, call
+from unittest.mock import call
 
 import pytest
 
-from app.main.s3_client import (
+from app.s3_client.s3_logo_client import (
     LOGO_LOCATION_STRUCTURE,
     TEMP_TAG,
     delete_temp_file,
     delete_temp_files_created_by,
     permanent_logo_name,
     persist_logo,
-    set_metadata_on_csv_upload,
     upload_logo,
 )
 
@@ -30,7 +29,7 @@ def upload_filename(fake_uuid):
 def test_upload_logo_calls_correct_args(client, mocker, fake_uuid, upload_filename):
     mocker.patch('uuid.uuid4', return_value=upload_id)
     mocker.patch.dict('flask.current_app.config', {'LOGO_UPLOAD_BUCKET_NAME': bucket})
-    mocked_s3_upload = mocker.patch('app.main.s3_client.utils_s3upload')
+    mocked_s3_upload = mocker.patch('app.s3_client.s3_logo_client.utils_s3upload')
 
     upload_logo(filename=filename, user_id=fake_uuid, filedata=data, region=region)
 
@@ -44,8 +43,8 @@ def test_upload_logo_calls_correct_args(client, mocker, fake_uuid, upload_filena
 
 def test_persist_logo(client, mocker, fake_uuid, upload_filename):
     mocker.patch.dict('flask.current_app.config', {'LOGO_UPLOAD_BUCKET_NAME': bucket})
-    mocked_get_s3_object = mocker.patch('app.main.s3_client.get_s3_object')
-    mocked_delete_s3_object = mocker.patch('app.main.s3_client.delete_s3_object')
+    mocked_get_s3_object = mocker.patch('app.s3_client.s3_logo_client.get_s3_object')
+    mocked_delete_s3_object = mocker.patch('app.s3_client.s3_logo_client.delete_s3_object')
 
     new_filename = permanent_logo_name(upload_filename, fake_uuid)
 
@@ -59,8 +58,8 @@ def test_persist_logo_returns_if_not_temp(client, mocker, fake_uuid):
     filename = 'logo.png'
     persist_logo(filename, filename)
 
-    mocked_get_s3_object = mocker.patch('app.main.s3_client.get_s3_object')
-    mocked_delete_s3_object = mocker.patch('app.main.s3_client.delete_s3_object')
+    mocked_get_s3_object = mocker.patch('app.s3_client.s3_logo_client.get_s3_object')
+    mocked_delete_s3_object = mocker.patch('app.s3_client.s3_logo_client.delete_s3_object')
 
     mocked_get_s3_object.assert_not_called()
     mocked_delete_s3_object.assert_not_called()
@@ -83,8 +82,8 @@ def test_delete_temp_files_created_by_user(client, mocker, fake_uuid):
     obj = namedtuple("obj", ["key"])
     objs = [obj(key='test1'), obj(key='test2')]
 
-    mocker.patch('app.main.s3_client.get_s3_objects_filter_by_prefix', return_value=objs)
-    mocked_delete_s3_object = mocker.patch('app.main.s3_client.delete_s3_object')
+    mocker.patch('app.s3_client.s3_logo_client.get_s3_objects_filter_by_prefix', return_value=objs)
+    mocked_delete_s3_object = mocker.patch('app.s3_client.s3_logo_client.delete_s3_object')
 
     delete_temp_files_created_by(fake_uuid)
 
@@ -94,7 +93,7 @@ def test_delete_temp_files_created_by_user(client, mocker, fake_uuid):
 
 
 def test_delete_single_temp_file(client, mocker, fake_uuid, upload_filename):
-    mocked_delete_s3_object = mocker.patch('app.main.s3_client.delete_s3_object')
+    mocked_delete_s3_object = mocker.patch('app.s3_client.s3_logo_client.delete_s3_object')
 
     delete_temp_file(upload_filename)
 
@@ -103,28 +102,10 @@ def test_delete_single_temp_file(client, mocker, fake_uuid, upload_filename):
 
 def test_does_not_delete_non_temp_file(client, mocker, fake_uuid):
     filename = 'logo.png'
-    mocked_delete_s3_object = mocker.patch('app.main.s3_client.delete_s3_object')
+    mocked_delete_s3_object = mocker.patch('app.s3_client.s3_logo_client.delete_s3_object')
 
     with pytest.raises(ValueError) as error:
         delete_temp_file(filename)
 
     assert mocked_delete_s3_object.called_with_args(filename)
     assert str(error.value) == 'Not a temp file: {}'.format(filename)
-
-
-def test_sets_metadata(client, mocker):
-    mocked_s3_object = Mock()
-    mocked_get_s3_object = mocker.patch(
-        'app.main.s3_client.get_csv_upload',
-        return_value=mocked_s3_object,
-    )
-
-    set_metadata_on_csv_upload('1234', '5678', foo='bar', baz=True)
-
-    mocked_get_s3_object.assert_called_once_with('1234', '5678')
-    mocked_s3_object.copy_from.assert_called_once_with(
-        CopySource='test-notifications-csv-upload/service-1234-notify/5678.csv',
-        Metadata={'baz': 'True', 'foo': 'bar'},
-        MetadataDirective='REPLACE',
-        ServerSideEncryption='AES256',
-    )
